Implements W3DAssetManager and GameAssetManager
Implements W3DAssetManager and GameAssetManager

A number of the virtual functions were found to not be in use and have
therefore been turned into assert(false) functions.
Free_Assets_With_Exclusion_List has not been implemented at this time.
It requires an additional class most likely and it is not clear if it
performs any useful function. Constructor of W3DAssetManager has not
been implemented at this time due to the requirement for multiple
PrototypeLoader classes. Non virtual member functions of
GameAssetManager require further work on MeshClass to implement.